MEDYAN Reference

Functions

Define a System

MEDYAN.StateDef

MEDYAN.SysDef

addreaction!

addreactioncallback!

addfilamentsite!

addfilament_reaction!

addfilamentendsite!

addfilamentend_reaction!

addlinker!

Mutate a Context

run_chemistry!

minimize_energy!

empty!

load_snapshot!

chem_newfilament!

chem_removefilament!

chem_polymerize!

chem_depolymerize!

chem_setmonomerstate!

chem_newlinker!

chem_removelinker!

chem_setlinkerstate!

chem_adddiffusingcount!

chem_addfixedcount!

adddiffusingcount_rand!

newfilament_rand!

set_time!

set_enable_cylinder_volume_exclusion!

set_mechboundary!

Get Context Info

filamentdata field

Interfaces

grids

Sites

filament sites

filament end sites

linker sites

linkable sites

States

linkers

Define a System

MEDYAN.StateDef — Type

The definitions and names for the state of a simulation

diffusingspeciesnames::Vector{Symbol}

fixedspeciesnames::Vector{Symbol}

filamentnames::Vector{Tuple{Symbol, Vector{Symbol}}}

linkernames::Vector{Symbol}

MEDYAN.SysDef — Type

A struct mapping names to id numbers in a simulation

statedef::MEDYAN.StateDef

diffusing::MEDYAN.PropDictionary

fixedspecies::MEDYAN.PropDictionary

filament::MEDYAN.PropDictionary

state::MEDYAN.PropDictionary

linker::MEDYAN.PropDictionary

linkerparams::MEDYAN.PropDictionary

filamentsite::MEDYAN.PropDictionary

filamentendsite::MEDYAN.PropDictionary

linkersite::MEDYAN.PropDictionary

linkablesite::MEDYAN.PropDictionary

allfixedspeciesnames::Vector{String}

compartmentreactions::Vector{MEDYAN.CompartmentReaction}

compartmentreactioncallbacks::Vector{Any}

MEDYAN.addreaction! — Function

addreaction!(s::SysDef,reaction::CompartmentReaction)::SysDef

addreaction!(s::SysDef,reactionexpr::AbstractString,rate::Float64,invvolumepower::Int)::SysDef

Add a reaction to the system. Return s

reactionexpr is a string describing the reaction stoichiometry

reactionexpr is comprised of reactant and product parts seperated by a "-->"

All whitespace characters are ignored.

Each side is then split by "+" to get the species names.

Repeated or extra "+" are ignored.

A species name can be prepended by a positive integer to represent multiple copies.

rate::Float64: Base rate for the reaction. ((nm³)^(invvolumepower)/s) rate constants correspond to stochastic rate constants in the sense used by Gillespie (J. Comp. Phys., 1976, 22 (4)).

invvolumepower::Int: volumefactor= (1/volume)^invvolumepower where volume is the volume of the compartment in nm³. Generally this is 0 for reactions without another diffusing reactant, and 1 if there is another diffusing reactant.

Example good reactionexpr

"diffusing.a + diffusing.b --> diffusing.c"
"diffusing.c --> diffusing.a + diffusing.b"
"+ + diffusing.c + --> + diffusing.a + + diffusing.b + +"
" --> diffusing.a + diffusing.b"
"diffusing.a + diffusing.b --> "
"diffusing.a + diffusing.a --> "
"2diffusing.a --> "
"2diffusing.a --> 20diffusing.a"
"diffusing.c + diffusing.b --> diffusing.c + diffusing.b"
"fixedspecies.rate1b --> fixedspecies.g"
"fixedspecies.rate1b + fixedspecies.g --> fixedspecies.g"
"fixedspecies.rate1b + 23fixedspecies.g --> fixedspecies.g"
"fixedspecies.g --> fixedspecies.rate1b + 23fixedspecies.g"
"fixedspecies.g + fixedspecies.rate1b--> 2fixedspecies.rate1b + 23fixedspecies.g"
"filamentsite.MT.d --> filamentsite.MT.d"
"filamentsite.MT.d + diffusing.a --> filamentsite.MT.d"
"fixedspecies.g --> diffusing.a"
"diffusing.a --> fixedspecies.g"
"filamentsite.actin.pm + diffusing.a --> filamentsite.actin.pm"

MEDYAN.addreactioncallback! — Function

addreactioncallback!(s::SysDef, reaction::CompartmentReaction, callback)::SysDef
addreactioncallback!(s::SysDef,reactionexpr::AbstractString,rate::Float64,invvolumepower::Int, callback)::SysDef

Like addreaction! but also adds callback. callback is called when the reaction happens with input of MEDYAN.Context and Int the compartment id where the reaction happened

MEDYAN.errorcheck_addcallback(callback,s::SysDef) can optionally be overloaded to add errorchecking when the callback is added.

MEDYAN.addfilament_reaction! — Function

Add filament reaction. Return s. Add a filamentsite and reaction with callback to change the monomer state. This can be used for filament aging, filament catalyzed reactions, or simple binding reactions.

Arguments

s::SysDef: the system to add to.

filamenttypename::Symbol: the filament type name.

filamentsitename::Symbol: the new name of the filamentsite added. This can be used as a catalyst in other reactions.

changedstatenames::Pair{Vector{Symbol}, Vector{Symbol}}: the changes to the monomer states, the first is the states to match. The second is the new monomer states after the reaction. both should be the same length. Ordered minus end first.

center::Int: Which index of changedstatenames.first is the actual location of the filamentsite. Used for determining what compartment the reaction goes in.

reactionexpr::AbstractString: Allows adding other reactants or products to the reaction. " + filamentsite.$(filamenttypename).$(filamentsitename) + " gets added to both sides this to create the full reaction expression. See addreaction! for syntax.

rate::Float64: Base rate for the reaction. ((nm³)^(invvolumepower)/s)

invvolumepower::Int: volumefactor= (1/volume)^invvolumepower where volume is the volume of the compartment in nm³. volumefactor only applies to this reaction not any other reaction using filamentsitename Generally this is 0 for reactions without another diffusing reactant, and 1 if there is another diffusing reactant.

Examples

using MEDYAN
statedef = StateDef(
	filamentnames= [(:filname,[
                            :a,
                            :b,
                            :c,
                        ]),
	],
)
s= SysDef(statedef)
addfilament_reaction!(s, :filname, :ab,
    [:a]=>[:b], 1,
    "-->", 1.75E-3, 0,
)
addfilament_reaction!(s, :filname, :aabc,
    [:a,:a]=>[:b,:c], 2,
    "-->", 1.75E-3, 0,
)

MEDYAN.addfilamentend_reaction! — Function

Add filament end reaction. Return s. Add a filamentendsite and reaction with callback to change the filaments. This can be used for polymerization, depolymeriation, and changing end state.

Arguments

s::SysDef: the system to add to.

filamenttypename::Symbol: the filament type name.

filamentendsitename::Symbol: the new name of the filamentendsite added. This can be used as a catalyst in other reactions.

isminusend::Bool: true if changing the minus end, false if changing the plus end.

changedendstatenames::Pair{Vector{Symbol}, Vector{Symbol}}: the changes to the end monomer states, the first is the states to match. The second is the new monomer states after the reaction. If the second has more states than the first, new monomers will be added, if the second has less, monomers will be removed. Ordered minus end first.

spacing::Float64: Space needed at the filament end for this reaction. (nm) ratefactor= exp(-β*spacing*loadforce) where β is 1/kT, loadforce is the external force pushing axially on the end of the filament. and ratefactor affects this reaction propensity and any others using filamentendsitename

reactionexpr::AbstractString: Allows adding other reactants or products to the reaction. " + filamentendsite.$(filamenttypename).$(filamentendsitename) + " gets added to both sides this to create the full reaction expression. See addreaction! for syntax.

rate::Float64: Base rate for the reaction. ((nm³)^(invvolumepower)/s)

invvolumepower::Int: volumefactor= (1/volume)^invvolumepower where volume is the volume of the compartment in nm³. volumefactor only applies to this reaction not any other reaction using filamentendsitename Generally this is 0 for reactions without another diffusing reactant, and 1 if there is another diffusing reactant.

Examples

using MEDYAN
statedef = StateDef(
	diffusingspeciesnames= [:a,],
	filamentnames= [(:filname,[
                            :plus,
                            :mid,
                            :minus,
                        ]),
	],
)
s= SysDef(statedef)
monomerspacing= 2.7
#minus end polymerization
addfilamentend_reaction!(s, :filname, :pm, true,
    [:minus]=>[:minus,:mid], monomerspacing,
    "diffusing.a -->", 10E3, 1,
)
#plus end depolymerization
addfilamentend_reaction!(s, :filname, :dpp, false,
    [:mid,:plus]=>[:plus], 0.0,
    "--> diffusing.a", 1.75E-3, 0,
)

MEDYAN.addlinker! — Function

addlinker!(s::SysDef,linkertypename::Symbol,defaultstate::LinkerState,mechparams;no_collide=false)

Add the linker to the system. Return s.

Setting no_collide to true allows cylinders this linker type connects and adjacent cylinders to clip.

```

Advanced Functions for Custom Site Types

MEDYAN.addfilamentendsite! — Function

addfilamentendsite!(s::SysDef,filamenttypename::Symbol,filamentendsitename::Symbol,site)::SysDef

MEDYAN.addfilamentsite! — Function

addfilamentsite!(s::SysDef,filamenttypename::Symbol,filamentsitename::Symbol,site)

Context

Mutate Context

The following functions should be used to mutate a Context so the Context remains valid.

deepcopy can be used to create an independent copy of a Context with the exception that they may share a random number generator.

These functions can also be used in a chemistry callback.

Core functions

MEDYAN.run_chemistry! — Function

run_chemistry!(c::Context, Δt)

Run chemistry for Δt time.

Update c.time.

MEDYAN.minimize_energy! — Function

minimize_energy!(c::Context)

Minimize mechanical energy of the context.

Base.empty! — Method

empty!(c::Context)

Remove all filaments, linkers, diffusing species, fixed species, and mechboundary.

MEDYAN.load_snapshot! — Function

load_snapshot!(c::Context,group::Union{HDF5.File, HDF5.Group})

Empty the context and load the state saved in the snapshot group.

c should be constructed with the same SysDef and parameters as the Context used to create the snapshot. Trying to load a snapshot into a Context constructed with a different system or parameters may lead to unexpected results.

See Snapshot group for more details.

MEDYAN.chem_newfilament! — Function

chem_newfilament!(c::Context; ftid=1, monomerstates, node_mids, nodepositions)

Return the filament id of a new filament.

Error if the filament isn't initially over 2 monomers long.

Newly added filaments don't have linkable sites, until after minimization.

Keyword Arguments

ftid=1: filament type id.

monomerstates: Collection of the MonomerState of the monomers in the new filament. In order from minus end to plus end.

length(monomerstates)>1

nodepositions: Collection of SVector{3,Float64}. The positions of the nodes, monomers are between nodes.

node_mids: Collection of Integer. The monomer ids at (slightly plus side of) the nodepositions

                               |
                    -----+-----|-----+-----
minus end <----       M  |  M  | (M) |  M        ----> plus end
                    -----+-----|-----+-----
                               |
                               ^ A nodeposition is indicated by the line.

The monomer id with parenthesis (M) will in node_mids

length(node_mids) == length(nodepositions) - 1

endloadforces=(0.0=>0.0): end load forces, usually don't use this, because load forces will automatically get updated at the next minimization.

MEDYAN.chem_removefilament! — Function

chem_removefilament!(c::Context; ftid=1, fid=lastindex(c.filamentdata[ftid]))

Remove the filament with type id ftid and id fid.

By default the last added filament with type id ftid is removed.

Warn if any monomers on the filament are referenced by any linkers, and remove those linkers.

The warning can be disabled by passing keyword argument warniflinkerremoved=false

MEDYAN.chem_polymerize! — Function

chem_polymerize!(c::Context, ftid, fid, isminusend::Bool, newstate::MonomerState)

Add a monomer with state newstate to the end of the filament with type id ftid and id fid.

If isminusend is true add the monomer to the minus end of the filament, if false add it to the plus end of the filament.

This doesn't affect the other monomer states, positions, linked linkers, or change any of the monomer names. New monomers are not linkable until after minimization.

MEDYAN.chem_depolymerize! — Function

chem_depolymerize!(c::Context, ftid, fid, isminusend::Bool)

Remove a monomer from the end of the filament with type id ftid and id fid.

If isminusend is true remove the monomer from the minus end of the filament, if false remove it from the plus end of the filament.

This doesn't affect the other monomer states, positions, linked linkers, or change any of the monomer names.

Error if the filament isn't initially over 2 monomers long.

Warn if the old end monomer is referenced in a linker, and remove the linker.

The warning can be disabled by passing keyword argument warniflinkerremoved=false

MEDYAN.chem_setmonomerstate! — Function

chem_setmonomerstate!(c::Context, name::MonomerName, state::MonomerState)

Update a monomer state.

MEDYAN.chem_newlinker! — Function

chem_newlinker!(c::Context, ltid, endnames::Pair{MonomerName,MonomerName}, linkerstate::LinkerState)

chem_newlinker!(c::Context, ltid, endnames::Pair{MonomerName,MonomerName};
    changedchemstate = (;),
    changedmechstate = (;),
    is_minimized = c.linkerdata[ltid].params.defaultstate.is_minimized,
)

Add a new linker with type id ltid between monemers endnames.

Error if a linker of the same type links the same monomers in the same order.

If the linker state isn't provided, c.linkerdata[ltid].params.defaultstate is new linker's state, optionally modified by the changedchemstate, changedmechstate, and is_minimized keyword arguments.

Keyword arguments changedchemstate and changedmechstate are namedtuples that shadow fields from the default chemstate and mechstate for the new linker.

Keyword argument is_minimized can be used to set the new linker's is_minimized state.

Example

chem_newlinker!(c, g.ltid, minusname=>plusname;
        changedmechstate = (L0 = dist,),
        changedchemstate = (numHeads = rand(g.numHeadsMin:g.numHeadsMax),),
    )

MEDYAN.chem_removelinker! — Function

chem_removelinker!(c::Context, ltid, endnames::Pair{MonomerName,MonomerName})

Remove a linker of type id ltid between monemers endnames.

Error if there is no linker there.

Don't modify monomer states.

MEDYAN.chem_setlinkerstate! — Function

chem_setlinkerstate!(c::Context, ltid, endnames::Pair{MonomerName,MonomerName}, linkerstate::LinkerState)

Set linkerstate of the linker of type id ltid with endnames.

Error if there is no linker there.

MEDYAN.chem_adddiffusingcount! — Function

chem_adddiffusingcount!(c::Context, sid, cid, inccount)

Add inccount to diffusing species id sid in compartment id cid

MEDYAN.chem_addfixedcount! — Function

chem_addfixedcount!(c::Context, sid, cid, inccount)

Add inccount to fixed species id sid in compartment id cid

MEDYAN.set_time! — Function

set_time!(c::MEDYAN.Context, x::Float64) -> Float64

Time (s)

MEDYAN.set_enable_cylinder_volume_exclusion! — Function

set_enable_cylinder_volume_exclusion!(c::MEDYAN.Context, x::Bool) -> Bool

Are cylinder volume exclusion forces calculated.

MEDYAN.set_mechboundary! — Function

set_mechboundary!(c::MEDYAN.Context; kwargs...) -> MEDYAN.MechBoundary

Keywords

planes::Vector{SVector{4,Float64}} = []

Planes that make up the mechanical boundary of the simulation

E = 1//2 * relu(pos ⋅ planes[bi][1:3] - planes[bi][4])^2

For example, a mech bounding plane [1,0,0,3] would try and make x < 3 nm with a spring constant of 1 pN/nm.

2.0*[1,0,0,3] would try and make x < 3 nm with a spring constant of 4 pN/nm.

capsules::Vector{SVector{8,Float64}} = []

Capsules that make up the mechanical boundary of the simulation.

capsules[bi][1:3] is the starting point of the spine line segment (nm). capsules[bi][4:6] is the axis of the spine line segment (nm). capsules[bi][7] is the radius (nm). capsules[bi][8] is the spring constant (pN/nm).

if capsules[bi][4:6] is zero then the capsule is a sphere. Capsule boundaries can be combined with plane boundaries to create cylinder boundaries.

E = 1//2*k*relu(d - r0)^2 where:

d is the distance of the point to the spine line segment described by capsules[bi][1:6].

r0 is capsules[bi][7]

k is capsules[bi][8]

Helpful extra functions

MEDYAN.adddiffusingcount_rand! — Function

adddiffusingcount_rand!(c::Context, dsid, inccount)

Distribute the added diffusing species count randomly to compartments weighted by volume.

dsid: diffusing species id.

inccount: amount to add.

MEDYAN.newfilament_rand! — Function

newfilament_rand!(c::Context, monomerstates; iterations = 10^9, ftid = 1)

Add a filament with type id ftid to the Context with random center position and direction.

Return the filament id of a new filament.

monomerstates is a collection of the MonomerState of the monomers in the new filament.

The filament will be inside the mech boundary.

Errors if it fails to add a filament.

The monomer are spaced by the value in the filament type's mechanical parameters.

Other functions to mutate a Context

filamentdata field

This is indexed by filament type id, filament id to yeild a MEDYAN.FilamentData struct.

A MEDYAN.FilamentData struct contains the position, orientation, and discrete state of each monomer in that filament and has the following getter methods.

MEDYAN.get_monomerstate — Function

get_monomerstate(f, mid)

Return the monomer state at monomer id mid in filament data f.

MEDYAN.firstmid — Function

firstmid(f)

Return the first monomer id of the filament data f.

MEDYAN.lastmid — Function

lastmid(f)

Return the last monomer id of the filament data f.

MEDYAN.get_monomerposition — Function

get_monomerposition(f, mid)

Return the monomer position at monomer id mid in filament data f.

MEDYAN.get_monomerplusvector — Function

get_monomerplusvector(f, mid)

Return the monomer plus unit vector at monomer id mid in filament data f. The plus vector points parallel to the filament axis towards the plus end.

MEDYAN.get_monomer_position_plusvector — Function

get_monomer_position_plusvector(f, mid)

Return tuple of get_monomerposition and get_monomerplusvector