Sites

Abstractly, sites are an interface that allow more complex structures like filaments to interact with diffusing species and compartment reaction propensities.

Each site type has an associated fixed species. During chemistry that fixed species count will always be the total counts of the site in the compartment.

Chemistry callbacks can try and pick a random site in a compartment to get specific information about exactly where in a compartment some event should happen.

These are the groups of sites currently in MEDYAN.jl:

filament sites

Filament sites can only be associated with one filament type.

Required Functions

A type T must implement the following functions to be a filament site.

getplusrange(site::T)::Int

How many monomer states in the plus direction this site looks at.

getminusrange(site::T)::Int

How many monomer states in the minus direction this site looks at.

filamentsitecount(site::T,states)::Float64

Return site count at a monomer given states, A vector of MonomerState with length length(states) == getminusrange(filamentsite) + 1 + getplusrange(filamentsite)

The site count gets put in the compartment of the center monomer.

Built-in filament sites

MEDYAN.FilamentSiteGeneral — Type

General filament site just matches monomer states

center::Int64

index of center monomer in states

states::Vector{UInt8}

monomer states to match

Related Functions and Fields

Context

MEDYAN.pickrandomfilamentsite function

Context.filamentsites field

SysDef

SysDef.filamentsite field

addfilamentsite! function

addfilament_reaction! function

filament end sites

Filament end sites can only be associated with one filament type.

Required Functions

A type T must implement the following functions to be a filament end site.

getrange(site::T)::Int

How many monomer states this site looks at.

isminusend(site::T)::Bool

True if the filament end site is on the minus end, false if on the plus end.

spacing(site::T)::Float64

Return spacing needed for the filament end site reaction to happen, units of nm.

filamentendsitecount(site::T,states)::Float64

Return the filament end site count at the end of a filament with monomer states vector states The rate of filament end reaction will be modified by the brownian ratchet model:

exp(-β*F*L) where F is the compression force on the filament end, and L is spacing(site).

length(states) == getrange(filamentendsite)

The site count gets put in the compartment of the end monomer.

Built-in filament sites

MEDYAN.FilamentEndSiteGeneral — Type

Filament end site that matches with a vector of monomer states.

isminusend::Bool

endstates::Vector{UInt8}

spacing::Float64

Related Functions and Fields

Context

MEDYAN.pickrandomfilamentendsite function

Context.filamentendsites field

SysDef

SysDef.filamentendsite field

addfilamentendsite! function

addfilamentend_reaction! function

linker sites

Linker sites can only be associated with one linker type. They can represent propensities for reactions to happen to a linker attached to two monomers, like unbinding or motor walking, or a site for a diffusing species to interact with a bound linker.

Required Functions

A type T must implement the following functions to be a linker site.

linkersitecount(site::T, linkerstate::LinkerState, mechparams, minusftid, plusftid, minusmonomerstates, plusmonomerstates, m_pos, p_pos, m_plusv, p_plusv)::Float64

Return the site count given the state of the linker, the minusftid, plusftid filament type ids, the minusmonomerstates, plusmonomerstates monomer states, and the positions and orientations of the two monomers.

minusmonomerstates and plusmonomerstates should be collections of 3 MonomerState ordered from minus to plus end, with the linker attached to the center monomer.

If a monomer doesn't exist, its state should be input as zero.

The site count gets put in the compartment of the minus end of the linker.

Built-in linker sites

MEDYAN.LinkerSiteOne — Type

Always returns 1.0

MEDYAN.LinkerSiteCount — Type

Returns linkerstate.chemstate.sitecount

MEDYAN.LinkerSiteSlipBond — Type

Models unbinding rates based on the following exponential form of Bell et al, 1978: k0*exp(f/f0) Where f is the magnitude of the force. So as to exponetially increase the unbinding with more force.

linkersitecount returns the unbinding rate in units of 1/s It returns k0 if the linker hasn't been minimized yet.

f0::Float64

The charicteristic force magnitude. Units of pN

k0::Float64

Unbinding rate at zero force. Units of 1/s

MEDYAN.LinkerSiteMotorCatch — Type

The catch-bond nature of myosin unbinding with multiple heads. Adopted from the results of Erdmann et al. 2013. The parallel cluster model.

As the force increases, the motor unbinding rate decreases to a minimum of k0/10

linkersitecount returns the unbinding rate in units of 1/s It returns k0 if the linker hasn't been minimized yet. It assumes linkerstate.chemstate.numHeads exists for the linker.

f0::Float64

single head characteristic unbinding force, units of pN/head

onRate::Float64

single head binding rate, units of 1/s

offRate::Float64

single head unbinding rate, units of 1/s

β::Float64

slope of head binding, units of (head)/(pN/head)

MEDYAN.LinkerSiteMotorStall — Type

Models myosin walking rates from the results of Erdmann et al. 2013. The parallel cluster model.

As the motor work per step increases, its walking rate goes to zero.

Returns the walking rate of one end. The callback can then sample which end to move, or reject the move.

fs::Float64

The stall force magnitude. Units of pN

k0::Float64

Walking rate of one end at zero force. Units of 1/s

α::Float64

Positive dimensionless parameter defining the steepness of the curve, smaller is more steep, if α is inf, the curve is linear

walking_direction::Int32

Motor walking direction, +1 is towards plus end, -1 is towards minus end

isminusend::Bool

which motor end does this represent

Related Functions and Fields

Context

MEDYAN.pickrandomlinkersite function

Context.linkersites field

SysDef

SysDef.linkersite field

addlinkersite! function

linkable sites

Linkable sites represent a potential link between two monomers, potentially on two different filaments. The site count is added to the compartment of the potential monomer that could attach to the minus end of the linker.

Required Functions

A type T must implement the following functions to be a linkable site.

cutoff_distance(site::T)::Float64

Return the maximum distance in nm of a linkable site.

in_linking_range(site::T,m_pos,p_pos,m_plusv,p_plusv)::Tuple{Bool,Bool}

Return if two monomers are in range to possibly link in the forward direction, and if they are in range to possibly link in the reverse direction.

getftids(site::T)::Tuple{Integer,Integer}

Return the filament types of the minus and plus monomers.

getmidsteps(site::T)::Tuple{Integer,Integer}

Return the monomer step size starting with mid 0 in the search for potential linkable sites on the minus and plus ends respectively.

1 searches every monomer. 2 searches every monomer with even mid ...

If the plus and minus filament types are the same, then getmidsteps(site)[1] == getmidsteps(site)[2]

linkablesitecount(site::T,minusstate,plusstate,m_pos,p_pos,m_plusv,p_plusv)::Float64

Return the site count of two monomers getting linked given their states, their positions, and their plus vectors.

Built-in linkable sites

MEDYAN.LinkableSiteRange — Type

Any pair within max to min range are linkable if they have matching state with minusstate and plusstate.

minusftid::Int64

plusftid::Int64

minusskip::Int64

plusskip::Int64

minusstate::UInt8

plusstate::UInt8

minrange::Float64

maxrange::Float64

MEDYAN.LinkableSiteMinAngleRange — Type

Any pair within max to min range are linkable if they have matching state with minusstate and plusstate and if both the angles between the filaments and line between the pair are greater than acos(cosminangle) for example if cosminangle is 1, all angles are ok, if cosminangle is sqrt(2)/2, and the line between the pair of monomers is parallel or anti parallel to either filament, it won't be linkable.

This can be used to prevent linkers from binding both ends to the same filament.

minusftid::Int64

plusftid::Int64

minusskip::Int64

plusskip::Int64

minusstate::UInt8

plusstate::UInt8

minrange::Float64

maxrange::Float64

cosminangle::Float64

Related Functions and Fields

Context

MEDYAN.pickrandomlinkablesite function

Context.linkablesites field

SysDef

SysDef.linkablesite field

addlinkablesite! function