MEDYAN Reference

Functions

Define a System

StateDefinition

SystemDefinitions

addreaction

addreactioncallback

addfilamentsite

addfilament_reaction

addfilamentendsite

addfilamentend_reaction

addlinker

Mutate a Context

chem_setmonomerstate!

chem_newfilament!

chem_newlinker!

Get Context Info

filamentdata field

Interfaces

grids

Sites

filament sites

filament end sites

linker sites

States

linkers

linker_force function

convert2mechdata function

Define a System

MEDYAN.StateDefinition — Type

The definitions and names for the state of a simulation

diffusingspeciesnames::Vector{Symbol}

compartmentspeciesnames::Vector{Symbol}

filamentnames::Vector{Tuple{Symbol, Vector{Symbol}}}

linkernames::Vector{Symbol}

MEDYAN.SystemDefinitions — Type

A struct mapping names to id numbers in a simulation

statedef::StateDefinition

diffusing::Any

compartmentspecies::Any

filament::Any

state::Any

linker::Any

linkerparams::Any

filamentsite::Any

filamentendsite::Any

linkersite::Any

linkablesite::Any

totalcompartmentspecies::Any

compartmentreactions::Any

compartmentreactioncallbacks::Any

Missing docstring.

Missing docstring for addreaction. Check Documenter's build log for details.

Missing docstring.

Missing docstring for addreactioncallback. Check Documenter's build log for details.

MEDYAN.addfilament_reaction — Function

Return a new SystemDefinition with added filament reaction. Add a filamentsite and reaction with callback to change the monomer state. This can be used for filament aging, filament catalyzed reactions, or simple binding reactions.

Arguments

s::SystemDefinitions: the system to add to.

filamenttypename::Symbol: the filament type name.

filamentsitename::Symbol: the new name of the filamentsite added. This can be used as a catalyst in other reactions.

changedstatenames::Pair{Vector{Symbol}, Vector{Symbol}}: the changes to the monomer states, the first is the states to match. The second is the new monomer states after the reaction. both should be the same length. Ordered minus end first.

center::Int: Which index of changedstatenames.first is the actual location of the filamentsite. Used for determining what compartment the reaction goes in.

reactionexpr::AbstractString: Allows adding other reactants or products to the reaction. " + filamentsite.$(filamenttypename).$(filamentsitename) + " gets added to both sides this to create the full reaction expression. See addreaction for syntax.

rate::Float64: Base rate for the reaction. ((nm³)^(invvolumepower)/s)

invvolumepower::Int: volumefactor= (1/volume)^invvolumepower where volume is the volume of the compartment in nm³. volumefactor only applies to this reaction not any other reaction using filamentsitename Generally this is 0 for reactions without another diffusing reactant, and 1 if there is another diffusing reactant.

Examples

using MEDYAN
statedef = StateDefinition(
	filamentnames= [(:filname,[
                            :a,
                            :b,
                            :c,
                        ]),
	],
)
s= SystemDefinitions(statedef)
s= addfilament_reaction(s, :filname, :ab,
    [:a]=>[:b], 1,
    "-->", 1.75E-3, 0,
)
s= addfilament_reaction(s, :filname, :aabc,
    [:a,:a]=>[:b,:c], 2,
    "-->", 1.75E-3, 0,
)

MEDYAN.addfilamentend_reaction — Function

Return a new SystemDefinition with added filament end reaction. Add a filamentendsite and reaction with callback to change the filaments. This can be used for polymerization, depolymeriation, and changing end state.

Arguments

s::SystemDefinitions: the system to add to.

filamenttypename::Symbol: the filament type name.

filamentendsitename::Symbol: the new name of the filamentendsite added. This can be used as a catalyst in other reactions.

isminusend::Bool: true if changing the minus end, false if changing the plus end.

changedendstatenames::Pair{Vector{Symbol}, Vector{Symbol}}: the changes to the end monomer states, the first is the states to match. The second is the new monomer states after the reaction. If the second has more states than the first, new monomers will be added, if the second has less, monomers will be removed. Ordered minus end first.

spacing::Float64: Space needed at the filament end for this reaction. (nm) ratefactor= exp(-β*spacing*loadforce) where β is 1/kT, loadforce is the external force pushing axially on the end of the filament. and ratefactor affects this reaction propensity and any others using filamentendsitename

reactionexpr::AbstractString: Allows adding other reactants or products to the reaction. " + filamentendsite.$(filamenttypename).$(filamentendsitename) + " gets added to both sides this to create the full reaction expression. See addreaction for syntax.

rate::Float64: Base rate for the reaction. ((nm³)^(invvolumepower)/s)

invvolumepower::Int: volumefactor= (1/volume)^invvolumepower where volume is the volume of the compartment in nm³. volumefactor only applies to this reaction not any other reaction using filamentendsitename Generally this is 0 for reactions without another diffusing reactant, and 1 if there is another diffusing reactant.

Examples

using MEDYAN
statedef = StateDefinition(
	diffusingspeciesnames= [:a,],
	filamentnames= [(:filname,[
                            :plus,
                            :mid,
                            :minus,
                        ]),
	],
)
s= SystemDefinitions(statedef)
monomerspacing= 2.7
#minus end polymerization
s= addfilamentend_reaction(s, :filname, :pm, true,
    [:minus]=>[:minus,:mid], monomerspacing,
    "diffusing.a -->", 10E3, 1,
)
#plus end depolymerization
s= addfilamentend_reaction(s, :filname, :dpp, false,
    [:mid,:plus]=>[:plus], 0.0,
    "--> diffusing.a", 1.75E-3, 0,
)

MEDYAN.addlinker — Function

addlinker(s::SystemDefinitions,linkertypename::Symbol,defaultstate::LinkerState,mechparams;no_collide=false)

Add the linker to the system. Returns a new system.

Setting no_collide to true allows cylinders this linker type connects and adjacent cylinders to clip.

```

Advanced Functions for Custom Site Types

Missing docstring.

Missing docstring for addfilamentendsite. Check Documenter's build log for details.

MEDYAN.addfilamentsite — Function

addfilamentsite(s::SystemDefinitions,filamenttypename::Symbol,filamentsitename::Symbol,site)

Return a new SystemDefinition with added filament site.

Context

Mutate Context

The following functions should be used to mutate a Context so the Context remains valid.

deepcopy can be used to create an independent copy of a Context with the exception that they may share a random number generator.

These functions can also be used in a chemistry callback.

MEDYAN.chem_newfilament! — Function

chem_newfilament!(c::Context; ftid=1, monomerstates, node_mids, nodepositions)

node_mids are the monomer ids to the right of each node,

nodepositions are the positions of the nodes, monomers are between nodes.

Return the fid

Newly added filaments don't have linkable sites, until after minimization.

MEDYAN.newrandomfilament! — Function

Add a filament to the Context with random center position and direction.

The filament will be inside all the bounding planes.

Errors if it fails to add a filament.

MEDYAN.chem_setmonomerstate! — Function

chem_setmonomerstate!(c::Context, name::MonomerName, state::MonomerState)

Update a monomer state.

Missing docstring.

Missing docstring for chem_polymerize!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for chem_depolymerize!. Check Documenter's build log for details.

MEDYAN.chem_newlinker! — Function

chem_newlinker!(c::Context, ltid, endnames::Pair{MonomerName,MonomerName}, linkerstate::LinkerState)

chem_newlinker!(c::Context, ltid, endnames::Pair{MonomerName,MonomerName};
    changedchemstate = (;),
    changedmechstate = (;),
    is_minimized = c.linkerdata[ltid].params.defaultstate.is_minimized,
)

Add a new linker with type id ltid between monemers endnames.

Error if a linker of the same type links the same monomers in the same order.

If the linker state isn't provided, c.linkerdata[ltid].params.defaultstate is new linker's state, optionally modified by the changedchemstate, changedmechstate, and is_minimized keyword arguments.

Keyword arguments changedchemstate and changedmechstate are namedtuples that shadow fields from the default chemstate and mechstate for the new linker.

Keyword argument is_minimized can be used to set the new linker's is_minimized state.

MEDYAN.chem_removelinker! — Function

chem_removelinker!(c::Context, ltid, endnames::Pair{MonomerName,MonomerName})

Remove a linker of type id ltid between monemers endnames.

Error if there is no linker there.

Don't modify monomer states.

MEDYAN.chem_setlinkerstate! — Function

chem_setlinkerstate!(c::Context, ltid, endnames::Pair{MonomerName,MonomerName}, linkerstate::LinkerState)

Sets linkerstate of a linker of type id ltid with endnames.

Missing docstring.

Missing docstring for setdiffusingspeciescount!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for adddiffusingspeciescount!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for adddiffusingspeciescount_random_compartments!. Check Documenter's build log for details.

Other functions to mutate a Context

Missing docstring.

Missing docstring for run_chemistry!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for minimize_energy!. Check Documenter's build log for details.

Base.empty! — Function

Remove all filaments, linkers, diffusing species, and compartment species.

Missing docstring.

Missing docstring for load_snapshot!. Check Documenter's build log for details.

filamentdata field

This is indexed by filament type id, filament id to yeild a MEDYAN.FilamentData struct.

A MEDYAN.FilamentData struct contains the position, orientation, and discrete state of each monomer in that filament and has the following getter methods.

MEDYAN.get_monomerstate — Function

get_monomerstate(f, mid)

Return the monomer state at monomer id mid in filament data f.

MEDYAN.firstmid — Function

firstmid(f)

Return the first monomer id of the filament data f.

MEDYAN.lastmid — Function

lastmid(f)

Return the last monomer id of the filament data f.

MEDYAN.get_monomerposition — Function

get_monomerposition(f, mid)

Return the monomer position at monomer id mid in filament data f.

MEDYAN.get_monomerplusvector — Function

get_monomerplusvector(f, mid)

Return the monomer plus unit vector at monomer id mid in filament data f. The plus vector points parallel to the filament axis towards the plus end.

MEDYAN.get_monomer_position_plusvector — Function

get_monomer_position_plusvector(f, mid)

Return tuple of get_monomerposition and get_monomerplusvector